refactor(api): move nmcli constants to own module
nmcli.py was getting a bit large, so let's move the big eap constants
into their own little place.
