+++ api/tests/opentrons/system/test_nmcli.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Tuple
 import pytest
 from pytest import MonkeyPatch
 
-from opentrons.system import nmcli
+from opentrons.system import network_constants, nmcli
 
 
 def test_parse_colonsep() -> None:
@@ -166,7 +166,7 @@ GENERAL.STATE:100 (connected)"""
     monkeypatch.setattr(nmcli, "_call", mock_call)
 
     assert await nmcli.is_connected() == "full"
-    assert await nmcli.iface_info(nmcli.NETWORK_IFACES.WIFI) == {
+    assert await nmcli.iface_info(network_constants.NETWORK_IFACES.WIFI) == {
         # test "--" gets mapped to None
         "ipAddress": None,
         "macAddress": "B8:27:EB:5F:A6:89",
@@ -176,7 +176,7 @@ GENERAL.STATE:100 (connected)"""
         "type": "wifi",
     }
 
-    assert await nmcli.iface_info(nmcli.NETWORK_IFACES.ETH_LL) == {
+    assert await nmcli.iface_info(network_constants.NETWORK_IFACES.ETH_LL) == {
         "ipAddress": "169.254.229.173/16",
         "macAddress": "B8:27:EB:39:C0:9A",
         # test missing output gets mapped to None
@@ -194,4 +194,4 @@ GENERAL.STATE:100 (connected)"""
     monkeypatch.setattr(nmcli, "_call", dummy_error_mock_call)
     assert await nmcli.is_connected() == "full"
     with pytest.raises(ValueError, match="this is a dummy error"):
-        await nmcli.iface_info(nmcli.NETWORK_IFACES.WIFI)
+        await nmcli.iface_info(network_constants.NETWORK_IFACES.WIFI)
